refactor(schemas): migrate Phase 33 endpoint schemas — 6 files, 12 classes (#6042)
Move local Pydantic class definitions from endpoint files into the matching
domain schema modules so they can be reused and discovered centrally.

- agent_config.py: AgentConfig, AgentModelUpdate → schemas_agent.py
- branch_health.py: BranchDivergenceResponse, BranchHealthResponse → schemas_system.py
- captcha.py: CaptchaResolutionRequest, CaptchaResolutionResponse → schemas_workflows.py
- elevation.py: ElevationRequest, ElevationAuthorization → schemas_workflows.py
- hot_reload.py: ReloadRequest, ReloadResponse → schemas_system.py
- analytics_embedding_patterns.py: EmbeddingUsageRequest (with helper methods),
  EmbeddingStatsResponse → schemas_analytics.py

Endpoint files now import from the schemas modules; no behavior change.
