+++ autobot-backend/api/analytics_embedding_patterns.py
@@ -27,8 +27,8 @@ from typing import Any, Dict, Optional
 
 from fastapi import APIRouter, Depends, HTTPException, Query
 from fastapi.responses import JSONResponse
-from pydantic import BaseModel, Field
 
+from api.schemas_analytics import EmbeddingUsageRequest
 from auth_middleware import check_admin_permission
 from autobot_shared.redis_client import RedisDatabase
 from autobot_shared.redis_mixin import AsyncRedisClientLockedMixin
@@ -129,71 +129,6 @@ class BatchOptimizationRecommendation:
 # =============================================================================
 # Pydantic Models
 # =============================================================================
-
-
-class EmbeddingUsageRequest(BaseModel):
-    """Request to record embedding usage"""
-
-    operation_type: str = Field(
-        default="document_vectorization", description="Type of embedding operation"
-    )
-    model: str = Field(..., description="Embedding model used")
-    provider: str = Field(default="ollama", description="Embedding provider")
-    token_count: int = Field(..., description="Total tokens processed")
-    document_count: int = Field(default=1, description="Number of documents processed")
-    batch_size: int = Field(default=1, description="Batch size used")
-    processing_time: float = Field(..., description="Processing time in seconds")
-    success: bool = Field(default=True, description="Whether operation succeeded")
-    source: Optional[str] = Field(None, description="Source of the operation")
-    metadata: Optional[Dict[str, Any]] = Field(None, description="Additional metadata")
-
-    # === Issue #372: Feature Envy Reduction Methods ===
-
-    def to_usage_record(self, operation_id: str, cost: float) -> Dict[str, Any]:
-        """Convert to usage record dict for storage (Issue #372)."""
-        return {
-            "operation_id": operation_id,
-            "operation_type": self.operation_type,
-            "model": self.model,
-            "provider": self.provider,
-            "token_count": self.token_count,
-            "document_count": self.document_count,
-            "batch_size": self.batch_size,
-            "processing_time": self.processing_time,
-            "success": self.success,
-            "timestamp": datetime.now(tz=timezone.utc).isoformat(),
-            "cost": cost,
-            "source": self.source or "unknown",
-            "metadata": self.metadata or {},
-        }
-
-    def get_tokens_per_second(self) -> float:
-        """Calculate tokens per second (Issue #372)."""
-        if self.processing_time > 0:
-            return self.token_count / self.processing_time
-        return 0
-
-    def get_log_summary(self) -> str:
-        """Get formatted log summary (Issue #372)."""
-        return (
-            f"{self.document_count} docs, {self.token_count} tokens, "
-            f"{self.processing_time:.3f}s"
-        )
-
-
-class EmbeddingStatsResponse(BaseModel):
-    """Response model for embedding statistics"""
-
-    total_operations: int
-    total_tokens: int
-    total_documents: int
-    total_cost: float
-    avg_processing_time: float
-    success_rate: float
-    avg_batch_size: float
-    tokens_per_second: float
-    period: str
-    timestamp: str
 
 
 # =============================================================================
