+++ autobot-backend/api/captcha.py
@@ -22,8 +22,8 @@ from typing import Optional
 
 from fastapi import APIRouter, HTTPException, Path
 from fastapi.responses import JSONResponse
-from pydantic import BaseModel
 
+from api.schemas_workflows import CaptchaResolutionRequest, CaptchaResolutionResponse
 from autobot_shared.time_utils import utc_timestamp
 from services.captcha_human_loop import CaptchaResolutionStatus, get_captcha_human_loop
 from api.schemas_common import DataResponse
@@ -31,22 +31,6 @@ from autobot_shared.error_boundaries import ErrorCategory, with_error_handling
 
 router = APIRouter(prefix="/captcha", tags=["captcha"])
 logger = logging.getLogger(__name__)
-
-
-class CaptchaResolutionRequest(BaseModel):
-    """Request model for CAPTCHA resolution"""
-
-    notes: Optional[str] = None
-
-
-class CaptchaResolutionResponse(BaseModel):
-    """Response model for CAPTCHA resolution"""
-
-    success: bool
-    captcha_id: str
-    status: str
-    message: str
-    timestamp: Optional[str] = None
 
 
 @router.post("/{captcha_id}/resolve", response_model=CaptchaResolutionResponse)
